refactor: file structure modified
