+++ src/app/Footer.tsx
@@ -1,4 +1,4 @@
-// components/Footer.tsx
+import React from "react";
 import Image from "next/image";
 
 const socialLinks = [
