+++ src/app/HeroBanner.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import Image from "next/image";
 
 export const HeroBanner = () => {
